requirements.txt
playwright==1.49.1
pydantic==2.10.3
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httpx==0.28.1
tenacity==9.0.0
python-dotenv==1.0.1
